tcp: Allow serial number to mismatch in reply (#197)
As it turns out the serial number _does not seem_ to be required in the
request packet, in order to extract data from the inverter.  This field
is present in the "header" around the TCP packet, most likely to keep
the format the same between requests and replies but is not actually
checked by the inverter before it sends back a valid reply.

By being more lenient here we can recommend - or at least request - the
community to test with a nonmatching (e.g. `0`) serial number, instead
of having to look it up on the screen or through the `javascript`
backend.