Decouple network request from device/inverter parsing
Currently both the `device()` and `inverter()` functions perform an
identical web request, but parse the data into a different `dataclass`.
This is not only unnecessarily slow and wasteful but also seems to act
up on certain inverters.

Remodel the code to return a wrapper object that encapsulates the reply
from the inverter, and parse it into a specific `dataclass` only when
the user asks for it.