Fix remainder of ruff lints (#365)
* Fix remainder of ruff lints

* Revert TCP error messages

General communication messages for read and write are caught and raised
during `wait_closed()`.

* Attempt to fix missing coverage on 108->exit

* Ignore `PGH003` for type skip

* use `type: ignore[unreachable]` instead

* Fix coverage by moving `self._close_session=False` out of `if`

Somehow `codecov` is not able to deduce that `_close_session` is `False`
if the `if` was not entered.

* Improve and reuse `close()` function

* Ignore type-less subclass with `misc` instead of `unreachable` ignore

The type is still reachable, `mypy` simply doesn't have type info hence
considers it as the `Any` type which shouldn't be subclassed:

    tests/test_tcp_models.py:206: error: Unused "type: ignore" comment
    tests/test_tcp_models.py:206: error: Class cannot subclass "TestCase" (has type "Any")  [misc]
    tests/test_tcp_models.py:206: note: Error code "misc" not covered by "type: ignore" comment

* Make content_type check less strict again