tcp: Make firmware strings (in packet suffix) optional
Multiple of the following reports are going around:

    Buffer size too small (85 instead of at least 125 bytes)

As it turns out these last 40 bytes contain exactly the `firmware` and
`firmware_slave` strings, which are not sent by certain 3000TL
inverters.  Move the fields to a secondary, optional `Structure` to
allow these devices to be read out over TCP, and warn when the length
of a reply isn't matching one of the known message sizes so that we can
potentially discover more different layouts.